+++ Pinta.Effects/Adjustments/SepiaEffect.cs
@@ -8,6 +8,7 @@
 /////////////////////////////////////////////////////////////////////////////////
 
 using System;
+using System.Threading.Tasks;
 using Cairo;
 using Pinta.Core;
 
@@ -23,7 +24,23 @@ public sealed class SepiaEffect : BaseEffect
 			ColorBgra.Black,
 			ColorBgra.White);
 
-	public SepiaEffect (IServiceProvider _) { }
+	private readonly IChromeService chrome;
+	private readonly IWorkspaceService workspace;
+	public SepiaEffect (IServiceProvider services)
+	{
+		chrome = services.GetService<IChromeService> ();
+		workspace = services.GetService<IWorkspaceService> ();
+
+		EffectData = new SepiaData ();
+	}
+
+	public override bool IsConfigurable
+		=> true;
+	public SepiaData Data
+		=> (SepiaData) EffectData!;
+
+	public override Task<bool> LaunchConfiguration ()
+		=> chrome.LaunchSimpleEffectDialog (this, workspace);
 
 	public sealed override bool IsTileable
 		=> true;
@@ -41,5 +58,20 @@ public sealed class SepiaEffect : BaseEffect
 	{
 		desaturate.Apply (destination, source, roi);
 		level.Apply (destination, destination, roi);
+		double strength = Data.Strength / 100d;
+		if (strength == 1d) return;
+		ReadOnlySpan<ColorBgra> sourceData = source.GetReadOnlyPixelData ();
+		Span<ColorBgra> destinationData = destination.GetPixelData ();
+		foreach (var pixel in Tiling.GeneratePixelOffsets (roi, source.GetSize ()))
+			destinationData[pixel.memoryOffset] = ColorBgra.Lerp (
+				sourceData[pixel.memoryOffset],
+				destinationData[pixel.memoryOffset],
+				strength);
+	}
+
+	public sealed class SepiaData : EffectData
+	{
+		[MinimumValue (0), MaximumValue (100)]
+		public int Strength { get; set; } = 100;
 	}
 }